open ddk files in DFPT part in xmpi_single mode as parallel IO is not needed

+++ tests/pymods/testsuite.py
@@ -1926,11 +1926,13 @@ pp_dirpath $ABI_PSPDIR
             if not os.path.exists(os.path.join(src, "README.md")):
                 self._status = "skipped"
                 msg = self.full_id + ": Skipped:\n\tThis test requires files in the git submodule:\n\t\t%s\n" % src
+                msg += "\tbut cannot find README.md file in dir\n"
                 msg += "\tUse:\n\t`git submodule init && git submodule update --recursive --remote`\n\t to fetch the last version from the remote url."
                 self.cprint(msg, status2txtcolor[self._status])
                 can_run = False
             else:
-                os.symlink(src, dst)
+                if not os.path.exists(dst):
+                    os.symlink(src, dst)
 
         self.run_etime = 0.0
